+++ site/plugins/admin/classes/SettingsManager.php
@@ -68,9 +68,23 @@ class SettingsManager
             $locales[basename($locale, '.yaml')] = $system_locales[basename($locale, '.yaml')];
         }
 
+        $entries = [];
+
+        foreach (Entries::getEntries('', 'date', 'DESC') as $entry) {
+            $entries[$entry['slug']] = $entry['title'];
+        }
+
+        $themes = [];
+
+        foreach (Filesystem::getDirList(PATH['themes']) as $theme) {
+            $themes[$theme] = $theme;
+        }
+
         Themes::view('admin/views/templates/system/settings/list')
                 ->assign('settings', Registry::get('settings'))
                 ->assign('locales', $locales)
+                ->assign('entries', $entries)
+                ->assign('themes', $themes)
                 ->display();
     }
 }