+++ site/plugins/admin/gulpfile.js
@@ -64,7 +64,9 @@ gulp.task('js', function(){
                    'node_modules/codemirror/mode/clike/clike.js',
                    'node_modules/codemirror/mode/yaml/yaml.js',
                    'node_modules/messenger-hubspot/build/js/messenger.min.js',
-                   'node_modules/messenger-hubspot/build/js/messenger-theme-flat.js'
+                   'node_modules/messenger-hubspot/build/js/messenger-theme-flat.js',
+                   'node_modules/clipboard/dist/clipboard.min.js',
+                   'node_modules/bs-custom-file-input/dist/bs-custom-file-input.min.js'
                 ])
     .pipe(sourcemaps.init())
     .pipe(concat('build.min.js'))