+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [0.8.2] - 2019-01-09
+### Added
+- Admin Panel: ClipboardJS added!
+- Admin Panel: Media Manager - Twitter Bootstrap File browser - added.
+- Admin Panel: Snippets Manager: Embeded code info modal added.
+- Admin Panel: Settings Manager - Select dropdown for default entry - added.
+- Admin Panel: Settings Manager - Select dropdown for timezones - added.
+- Admin Panel: Settings Manager - Select dropdown for themes - added.
+
+### Changed
+- Flextype Core: Text Components updated to 1.1.2.
+- Admin Panel: Entries Manager - upload file size increased to 27000000 bytes.
+
+### Fixed
+- Flextype Core: Default theme - issue with package.json - fixed.
+- Flextype Core: Code fixes according to the phpDocumentor.
+- Flextype Core: Default theme - settings and manifest - fixes.
+- Admin Panel: Translates fixes.
+
 ## [0.8.1] - 2019-01-07
 ### Added
 - Flextype Core: Glide/Intervention Image Implemented for Image manipulation!